server: maintain list of servers (allow duplicate elimination)
Because people tend to use the same servers in multiple clusters (and
actually by mistake in the same cluster) keep a list so that we can do
duplicate elimination.  Side effect of this is that it is no longer
necessary for the router to produce a list of servers in use.