server_queuereader: use queue_putback() on failures
Putting entries we retrieved (but failed to be sent) at the end of the
queue, actually changes the order and might cause expiration of more
recently submitted entries before the re-submitted one, which is
slightly bad.  Hence, if there is space, put back the entry at the front
of the queue, such that it is read and expired first.