server: keep connections open if we're writing a lot
Writing data to a closed socket (remote) will succeed.  The next write
will raise an error as the underlying subsystem has detected the
previous write failed.  This means we lose a metric without us noticing.
For that reason we initially reconnected for each batch we did so we
tried reconnecting fairly often and see the failure at that stage.
Downside of that is that we open a lot of sockets, which keep lingering
around a while on servers.  Hence, some hybrid solution, keep the
connection alive as long as we have work.  As soon as we're idle,
disconnect.