server_queuerunner: claim a bunch of metrics to write
It feels kind of pointless, but dequeue a whole bunch at a time in a
local "queue" to send it.  Advantage of the local queue is that we don't
have to push back into the queue, and don't have locking overhead, since
we're the only one reading from it.  So overall, this improves
performance at the expense of some memory.