router: change any_of routing to be more deterministic
Instead of letting threads decide who picks up what part of a shared
queue, distribute the incoming data among the members in a consistent
fashion (based on a hash of the metric).  This allows for example to
use any_of in front of a cluster of stores on the same machine, where
the unavailability of one of the stores doesn't mean the events won't
get written any more (like carbon_ch would do).