relay: implemented fnv1a_ch cluster type
Renamed carbon-hash to consistent-hash, to make it more general, and
suitable for reuse.  The fnv1a_ch cluster type makes directly use of
this oportunity by implementing a consistent hash cluster but using a
cheaper hash algoritm (FNV1a, also used in any_of) that is also defined
to take input from both host and port for each server.  This allows to
use a consistent hash strategy for a single hosts but multiple ports,
such as in issue #5.