server: redid any_of implementation
Instead of picking from other servers that seem to need help when
idling, changed the approach to actively pushing away metrics when in
need for it.

With the previous approach we'd easily see massive dataloss for a failed
target in the any_of cluster when the cluster received a fair amount of
metrics (since all the others were too busy to help out).  Therefore,
this approach should try harder not to lose any data, by actively moving
metrics to queues of the others.