dispatch: implemented disconnect of idle clients
It seems that some clients tend to stay connected whilst not doing
anything.  This is a problem for applications that assume the server
will close them (Windows) because a relay typically has many more
clients connected to it, so it is running out of sockets sooner than the
client in that case.
Disconnect idle clients after 10 seconds, so we're well beyond the
disconnect time of the server (in case of another carbon-c-relay).