dispatcher: increase client idle timeout to 10 minutes
When we disconnect, a connected client will not notice it is
disconnected until after it does a write (or performs a read, but that's
very unlikely to happen).  Hence if we disconnect too soon for clients
that keep the connection open for a while, initial metrics will get
lost.  We cannot solve this problem, but we /can/ try to make it happen
less by increasing the idle timeout.  Since most applications write
minutely stats, taking a 10-fold of that should barely affect any
application, except those applications we introduced the timeout for
(which seem to open up a connection and forget about it afterwards).