dispatch_connection: also accept tabs as separators
Patch based on
https://github.com/pkittenis/carbon-c-relay/commit/19a3a92c919cf373b853ad7a2faad312b427d7d3
by Panos:

Graphite accepts both spaces and tabs as separators in line metric
format but carbon-c-relay only accepts spaces. This causes relay to drop
line metric messages that would otherwise be accepted by graphite and
are sent by some clients.

Sensu for example uses tabs when sending data to graphite and these
metrics are not accepted by carbon-c-relay but are accepted by carbon
caches and the built in carbon relay.

Graphite will actually accept multiples of spaces and or tabs as
separators - have not seen clients that use multiple whitespace in that
way though.

---

In contrast to the original patch, this patch will make the relay accept
a mix of tabs and spaces and fold them into one space, for consumers
that require that as we found out in the past.  Also, with this patch,
carbon-c-relay will never write tabs, but convert everything to spaces.