collector: get timing right
We actually count microseconds (usec) not milli or nanoseconds.  Because
fixing the metric data would mean a multiply by 1000 (and not getting
any higher precision), I chose to change the metric names to reflect the
correct granularity, since the old metrics are sort of bogus anyway.

This patch is based on the work of ianslai in
https://github.com/ianslai/carbon-c-relay/commit/c9b46ae90661e170b30c1ebdb7486a98b9bb8e58