router: fix multi-threading corruption for rewrites
Due to a thinko, the pmatch holding space was put in the routing rule,
which can (and will be) be evaluated by multiple dispatchers at the same
time, causing concurrent (over)writes to the same buffer with undesired
rewrites as result.  Net effect is seemingly garbled metrics, since
different offsets are taken from the input string to produce the new
metric.