relay: only report dropped metrics in debug mode
We recently DOSsed ourself by just having a broken server as target.
Due to the output of metrics being dropped (which was logged to disk),
we trashed the whole relay getting stuck on writing data.  Lesson
learnt: don't do this.  Since it can be useful in certain scenarios,
allow to enable it via debug mode.