dispatch_connection: make sure we count all metrics received
In commit 30aa7049 the counting was moved down to after all servers
received the metric, however, in case of aggregation, no servers are
being sent to so they would never be counted.
Moved the counting to right after we determined the metric is OK, such
that we count all incoming metrics properly.