aggregator: make splay invisible to the consumer (other than delayed)
For issue #72, changed the way splay is added to the aggregations to
avoid a thundering herd when expiries happen.  Instead of changing the
bucket start-time, now the expiry time is changed (made longer) to
achieve the same thing, but metrics sent at the same timestamp will fall
in the same time-interval buckets, thus being aligned, which is
necessary for some computations.