remove FILELOGIP
The src ip is not accessible at the router level, we would have to pass
it on, but also figure out what it would be for we accept TCP/UDP/UNIX
connections and don't destinguish them in the dispatcher.  Lot of work,
probably very useful, but I'm not ready for that yet here and now.