dispatch_connection: feed correct dests size to router_route
Bad me, sizeof() gives size in bytes, not the size of the array.  This
allows for overrunning the dests array quite a lot, with all kinds of
interesting results.  I'm not sure if this is the cause for issue #62,
but it is for sure a problem for configs that send metrics to many
targets (dests is only 32 long).