entrycmp: compare to ip, port and/or instance
For issue #84, extend the comparator to match original fields if a
position collision would occur.  This commit is similar to
pull request #85, but mostly differs in the implementation of entrycmp
for cases where the instance is unset, or the hash-type is fnv1a.

To allow to revert behaviour the code can be compiled with
   -DCH_CMP_V40_BEHAVIOUR
in CFLAGS.  This also allows to compare the hash dumps.  For simple
tests I've done so far, no changes occurred (e.g. due to no collissions
in the original map).