aggregate: add send to cluster(s) support
This code works via an intermediate stub route to get the aggregation
results sent to the desired clusters.  This shouldn't affect performance
much, since the only way to route traffic out of the aggregators before
this change was in the very same way.  The only difference here, is that
the stubs use a unique prefix that is removed like rewrite, but
non-permanent, such that matching can occur without affecting the
traffic stream.

This code is very untested, next commits will add some docu and possibly
some bugfixes.