ch_get_nodes: fix jump hash metric submission
Fix incredibly stupid bug mainly caused by the fact my fnv1a
implementations are macros, not functions.  Anyway, metric was
incremented up to firstspace to calculate the hash, so strdupping it
afterwards is not a good idea.  Do the walk with a temp-var, as the
macro suggests to not tamper with the original metric pointer.