relay: revise reload to handle aggregations
For issues #32 and #133, get the aggregator to expire as if it was
shutting down on a reload of the configuration.  Side effect of this is
that we need to suspend the dispatchers such that we don't get data
while we're clearing out the aggregators.
In this scheme, a reload will cause a slight service interruption.
New connections are still accepted, but no data is read while the
aggregations are flushed, and the dispatchers load the new
configuration.  The upside of this is that there will be an atomic
reload, that is, everything is either routed according to the old
configuration, or the new configuration, not both at the same time as
was the case before this commit.