router_printconfig: do not show send to for aggregators in debug mode
Since in debug mode we print the real routing setup (with a match that
does a trick under the hood), it is confusing (and wrong) to print the
send to for the aggregation as well.  By hiding it, we print the real
routing performed, instead of the conceptual routing that is easier to
understand.