router_test_intern: fix segfault on aggregation without rewrites
Fix for issue #135.  When the rewrite wasn't called, a direct
replacement was written, but newfirstspace wasn't set.  Make sure it is
set, and while we're at it, fix a bug where the stub would be taken into
account if it /were/ a rewrite.

Unfortunately I couldn't use the pull request from the issue, since that
commit introduced unrelated behaviour changes and also it wasn't clear
what the suggested fix was.  It seemed to add a space somewhere without
fixing the issue for real.