reload: get aggregations to be sent before reloading config
As continuation for fixing issue #133, add substantial amount of code to
be able to achieve sending expired aggregations when reloading the
config.  We need this to make sure that aggregates that are sent as part
of the shutdown are routed with the configuration they were created for.
This is in particular necessary for metrics that hold pseudo stubs to
perform target routing.

So now we hold all dispatchers, then take the first and make it run the
internal_submission server specifically (where aggregations are being
written to) and drain that queue.  After that we reload the
configuration for all dispatchers and continue serving as normal.