server_queuereader: set write timeout to something patient
When the relay temporarily can't handle the output traffic, it will
(attempt to) stall clients, more or less asking for their patience.
Therefore, it is a good idea to, when the relay is a client itself, be
patient instead of reconnecting almost immediately.  Take a write
timeout of 10s or so, so we can detect hangs timely, but also wait a bit
for the upstream consumer when it wants us to.