collector: fix corrupted metrics for file clusters
This commit fixes two aspects of issue #159.  The first strncpy doesn't
overrun the buffer, but doesn't add trailing \0 either.  This caused the
"corrupted" metrics, because garbage was read, so use plain snprintf
instead to avoid that.  Second, 32 bytes (as originally intended for
IPv4 address) is a bit short, and insufficient for IPv6 addresses or
paths, so enlarge the buffer size so most paths will fit.