relay: introduce trace mode
This mode is only available when compiled with -DENABLE_TRACE because it
will have an impact on performance even when not enabled at runtime.
Use -dd or -ddd to enable trace, or trace and debug.  Currently only
shows metrics being handled by dispatchers, but is meant to be extended
in the future upon need for deep debugging the relay.