Add route using option to match
The purpose is to perform a temporary modification on the key so
that all of the appropriate keys end up on the same aggregation
instance.

This re-uses the rewrite methods to perform the modification of the
key, so the same functionality exists.  This does not affect the
metric key sent downstream.  It only affects the key used to
perform the routing.

Sample configuration would be:
cluster carbon
  carbon_ch
    10.0.0.1:2003=a
    10.0.0.1:2004=b
    10.0.0.1:2005=c
    10.0.0.1:2006=d
    10.0.0.2:2003=a
    10.0.0.2:2004=b
    10.0.0.2:2005=c
    10.0.0.2:2006=d

match ^collection\.([^.]+)\.([^.]+)\.([^.]+)\.[^.]+\.(.+)
  route using collection.\1.\2.\3.\4
  send to carbon
  stop
  ;

match *
  send to carbon
  stop
  ;