router_readconfig: properly set last pointers after include
Fix for issue #240, we didn't properly set the last pointers, causing
values to be overridden, rendering them invisible.  Rewrite rules add
routes and clusters, which in the issue are the done when nothing else
was added before, e.g. we incorrectly didn't re-read ret members.  Since
NULL values can be turned (like in this case) a re-traversal is
necessary.