dispatch_connection: treat UDP packets as fully contained
As reported in issue #241, the relay was concatenating UDP packets until
it found a newline to terminate a line.  Since UDP packets don't
necessarily arrive (or in order), one can't rely on sending partial
fragments.  So instead, make sure we treat each received UDP packet as
a full message (and all we're going to get).