dispatch_connection: don't wrongly assume noexpire == udp
Both aggregator and UDP connections have the noexpire bit set, but we
don't want to treat aggregator connections as all-in-one packet
connections.
So, instead of messing around with len (loosing the ability to say
something about whether work was done or not) and making some implicit
assumptions, introduce an isudp bit, which we can use instead to trigger
the incomplete message discard (6a17865).

This is likely the fix for issue #242, because it uses aggregations, and
the aggregator gets heavily blocked since all work the dispatchers did
were flagged as "no-work" leading to sleeps.