server_queuereader: only retry interrupted and incomplete writes
Incomplete writes should actually not occur, since we use blocking IO
here, but reality has shown that it does happen.  When we find an error,
only retry if it is an interruption.  In either case, when retrying
pause for a short while to allow catching up of the remote side.