dispatch_connection: always convert NULL-characters into _
Thanks to Piotr Popieluch for finding the cause of issue #206.  This
commit is much like his patch, but moves the NULL-check up so also
values/timestamps can't result in incomplete metrics like this.

Because the relay treated \0 as a normal character (due to strchr
finding a match at the end of its haystack string) the string being
copied (queued) for sending didn't include the full statement that was
validated.  Most notably, no newline.  This resulted in a partial write
when sending the metric out, which was observed by the sender as a
metric prefixed with some random garbage.