conffile: first step towards a flex/bison parser
Because of weaknesses in the current parsing, and complexity of soon to
be added parsing constructs, this is a work in progress to change the
config file parsing to a flex/bison-based parser.

In this checkpoint commit the skeleton work for the lexer (flex) is
done, the parser (bison) recognises cluster structures, and nothing
more.  Error reporting is setup as well.  The router just terminates
after parsing, e.g. this commit makes this a compiling, but totally
non-functioning router.