conffile: add lots of state to the lexer
To allow identifiers (most notably cluster names) to have names matching
keywords that aren't in use (log, send, ...) restrict the lexer to
tokens that are expected according to the grammar.  This is slightly
double, and the quoting syntax is there to solve this, but for backwards
compatibility we have to do this (for now).