relay: add statistics clause
The statistics clause is more extended than the send statistics to
clause and allows to set collector interval, summary mode, statistics
prefix and statistics destination from the config.

Issue #249, issue #113 and issue #25 could use this to emit the
statistics exactly as they want without rewrite rule and directly to the
destination of their choice.