dispatch_process_dests: never block the aggregator expiry thread
This is an attempt to do something about issue #250, but may be not
desirable, because bursts generated by the aggregator will likely result
in drops if the queue pressure is high.  On the other hand, we try to
spread the aggregation load, and server queues should be adjusted to
handle this in the normal case.
Idea is that blocking the expiry thread might cause an avalanche of
blocks resulting in a spiraling behaviour towards a full stop.