consistent-hash: perform binary search iso forward scan
Since the parser was redone, we know in advance how many servers we're
going to add to the ring.  This information we can use to allocate the
right amount of buckets for the ring during creation, which in turn
allows to have a single array containing the bucket entries --
decomplicating the whole administration a lot.  And it is much easier to
do binary search on it instead of pointer chasing.

This started out as using pull request #236, but because of the
opportunity created by the new parser, went a different route, aiming
for less memory consumption, but still much less lookup cost.  This
implementation is a compromise between the original and the single point
lookup one as suggested in the pull request.