conffile: return error on failing included files
The crash observed in issue #263 was due to parsing continuing with the
next file in a globbed include, which tried to work with an invalidated
and freed environment.  Instead, make sure we return an error if an
included file errored, such that we stop reading any pending files and
avoid this crash.