set TCP_USER_TIMEOUT to 10 seconds
This sets the TCP_USER_TIMEOUT to 10 seconds on outgoing connections.

When for some reason no ACK's are being received on an ESTABLISHED connection
the kernel will start retransmitting. On my system (Fedora 25) it takes about
15 minutes before the kernel decides to give up and return an error. This
reduces this time to 10 seconds so the relay can close the connection and open
a new one.