relay: revise when to submit sockbufsize warning
For issue #287, only emit warning about OS rejecting socket bufsize when
the value is actually lower than requested.  On Linux it seems the
kernel doubles the value (so it will never match), on other kernels it
seems not to, but might align or something.  So with this in place on
Linux the kernel might still set it to a lower value (because it
doubles), but solving that needs kernel-specific checks which I want to
defer for now.