router: ensure route using doesn't modify metric, PR #282
This is an alternative take to the original PR.  Instead of having
ch_get_nodes copy the metric into the destinations struct
(undocumented), perform the copy in router_route_intern, like for all
other cases.  This is more transparent, and ensures that the correct
metric is added to the send queue.