dispatch_process_dests: fix off-by-one error
Ensure we do not memmove anything when we succeed sending metrics to all
destinations.  Not that this changes much, because memmove 0 bytes is
fine, it returns success instead of failure.  In dispatch_connection
returning failure resulted in a less efficient/stalled code-path, so
this change should improve performance quite a bit.