dispatcher: call inflate() manually iso gzread(), issue #303
The gzread() implementation doesn't deal with non-blocking IO, while the
relay relies on it heavily.  Wrap inflate around a read buffer and call
inflate such that partial reads are handled fine.  An extra bonus here
is that results are returned sooner because we can instruct inflate to
return decompressed data ASAP.