consistent-hash: ensure get_nodes returns the first pos >= search
When collisions occur, we would stable-sort them such that the ring
would always be the same, regardless of input order.  However, the
binary search method (historical mistake) could end up on a dupicate
pos, and take the server as response, clearly not honouring the contract
of returning the /first/ >= pos match.
This change ensures collisions on pos are voided, and basically restores
pre-binary-search distribution introduced in v3.1.

This change should match the ring output with what bucky expects
https://github.com/jjneely/buckytools/issues/17