dispatcher: move listeners initialisation to dispatch_new_listener
When only udp connections were added, the listeners array was never
initialised.  Initialise the array now unconditionally before a listener
is added, such that the array is there.  We better not try to avoid
starting the listener for this case, as a future configuration loaded
via SIGHUP might add a tcp/unix listener again.

This fixes the crash reported in issue #318.