regex: allow usage of onigurama, pcre2 or pcre over libc implementation
The usage of an external library greatly speeds up in certain scenarios,
see: http://lh3lh3.users.sourceforge.net/reb.shtml
From there, it shows that onigurama in general gets the best results,
and in the worst case (alternatives) still is faster than libc.  Hence
onigurama is the preferred default when it is available.