router_optimise: break infinite loops, fix misoptimisation
Added a guard when seq goes out of order with the found groups sequence
number such that we don't go into an infinite loop.
Fixed the scenario where groups need to be terminated due to stop rules,
such that sequence is correctly incremented, and the group isn't
searched in anymore.

Closes: https://github.com/grobian/carbon-c-relay/issues/357