relay: add support for TLS min and max proto versions #442
Add parser support for protomin <X> and protomax <X> where X is one of:
  ssl3 tls1.{0,1,2,3} (currently supported variants from OpenSSL).

The versions are fed to openssl's config, which will deal with it from
there.  I used protomax/min because that's the preferred way of doing so
according to recent OpenSSL manpages.

Issue: https://github.com/grobian/carbon-c-relay/issues/442
Signed-off-by: Fabian Groffen <grobian@gentoo.org>