collector: fix behaviour regarding stdio streams when submission + debug
Our tests use -s -d to enable submission mode and debug.  This
previously should enable iostat-like output from the relay, but due to
stdio buffering on most libcs, this was invisible (a bug).  Using musl
libc, a different buffering strategy reveals we were actually writing
the statistics header and data, which for tests is undesirable.

So, rework the iostat-like behaviour into a new -S option, that implies
submission mode (-s).  While at it, cleanup a lot of cruft from old mode
setting that is no longer in use.

Bug: https://bugs.gentoo.org/833846
Signed-off-by: Fabian Groffen <grobian@gentoo.org>