router: don't route metrics to a failed any_of destination, #454
Make sure we respect the failed state of destinations so we don't
introduce artificial lag, and possibly an avalanche situation multiple
times which causes metric drops.
When an any_of destination fails and after a brief period of it not
returning, the destination is flagged as failed, and other destinations
from the any_of cluster will receive the metrics.  In the meanwhile the
queue from the now failed destination will be offloaded to other
destinations (causing a spike of traffic).

Closes: https://github.com/grobian/carbon-c-relay/issues/454
Signed-off-by: Fabian Groffen <grobian@gentoo.org>