faketime: use regular symbol iso nasty injection approach
macOS doesn't allow injecting symbols (any more) and it's fugly as well,
so instead of preloading, just compile a separate binary for testing
purposes using the time/gettimeofday functions from faketime.
This is activating using a define via build, and therefore function nor
override are available in the produced binary (which is good).

Signed-off-by: Fabian Groffen <grobian@gentoo.org>