aggregator: reduce splay to just a bunch of seconds
When large expiry windows are used, using a splay of up to a window is
just overkill.  We just want to avoid emptying all buckets at once, but
in the end, a little splay here should be enough.  If we really need to
stretch it out over the entire expiry windows, it means we're at max
capacity, and any random mishap, may go over the limit.

Closes: https://github.com/grobian/carbon-c-relay/issues/435
Signed-off-by: Fabian Groffen <grobian@gentoo.org>