load whole ssl certificate chain when loading certificates
This is needed for proper supply of whole certificate chain. Otherwise
intermediate certificates from pem file are not loaded and served which
causes SSL validation errors.

Even recommended by openssl documetation:
```
SSL_CTX_use_certificate_chain_file() should be used instead of the SSL_CTX_use_certificate_file() function in order to allow the use of complete certificate chains even when no trusted CA storage is used or when the CA issuing the certificate shall not be added to the trusted CA storage.
```

Signed-off-by: Adam Tkac <vonsch@gmail.com>