router_rewrite_metric: add optimisation for expansionless replacement
In issue #465 the observation was made that different behaviour was seen
based on a replacement with or without expansions.  The reason for that
is a shortcut the code made to avoid expensive replacement code.
Problem, however, was that that shortcut doesn't consider replacing the
matched expression only, thus retaining parts of the input metric
surrounding the matched part.

Add a shortcut to not do the expensive bit of expansions, when none are
to be applied in router_rewrite_metric, such that we can get the exact
same retention of the metric parts surrounding the match.  This makes
the behaviour consistent in both cases.

Issue: https://github.com/grobian/carbon-c-relay/issues/465
Signed-off-by: Fabian Groffen <grobian@gentoo.org>