router: fixup rewriting to be consistent
As noted in issue 465 after the initial fix from aca4289, the behaviour
seemed odd at best for some expressions.  Turns out that when no capture
groups are used, regcomp does not return any match range information.
Yet in order to consistently replace matching parts of the input metric,
we need this range.  Force the range information to be present by
wrapping the entire input expression in a capture group, if there are
none defined.  This way replacement of e.g. `.*` matches the whole
input, and thus replaces the whole input (greedy matching, as is
default), while `b.*` would replace everything from the first b until
the end of the metric.

Closes: https://github.com/grobian/carbon-c-relay/issues/465
Signed-off-by: Fabian Groffen <grobian@gentoo.org>